raw_data_postprocess/raw_data_postprocess/requirements_ec2.txt
requests
amazon-textract-response-parser==1.0.2
boto3==1.34.39
botocore==1.34.39
jmespath==1.0.1
marshmallow==3.20.2
numpy==1.26.4
packaging==23.2
pandas==2.2.0
Pint==0.23
python-dateutil==2.8.2
pytz==2024.1
rapidfuzz==3.6.1
s3transfer==0.10.0
six==1.16.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.0.7
python-dotenv==1.1.0